Improve get_compiled_contract_class_async
